+++ src/main/java/gift/kakao/service/KakaoOAuthService.java
@@ -2,15 +2,12 @@ package gift.kakao.service;
 
 import gift.kakao.dto.KakaoUserProfile;
 import jakarta.servlet.http.HttpSession;
-import java.time.Duration;
 import java.util.Map;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.MediaType;
-import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestClient;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.context.request.RequestContextHolder;
@@ -18,7 +15,7 @@ import org.springframework.web.context.request.ServletRequestAttributes;
 import org.springframework.web.util.UriComponentsBuilder;
 
 @Service
-public class KakaoOAuthService implements OAuthService, KakaoMessageService {
+public class KakaoOAuthService implements OAuthService {
 
     @Value("${kakao.client-id}")
     private String clientId;
@@ -34,15 +31,8 @@ public class KakaoOAuthService implements OAuthService, KakaoMessageService {
 
     private final RestClient restClient;
 
-    public KakaoOAuthService(RestClient.Builder restClientBuilder) {
-        SimpleClientHttpRequestFactory requestFactory = new SimpleClientHttpRequestFactory();
-        requestFactory.setConnectTimeout((int) Duration.ofSeconds(5).toMillis());
-        requestFactory.setReadTimeout((int) Duration.ofSeconds(3).toMillis());
-
-        this.restClient = restClientBuilder
-            .baseUrl("https://kauth.kakao.com")
-            .requestFactory(requestFactory)
-            .build();
+    public KakaoOAuthService(RestClient restClient) {
+        this.restClient = restClient;
     }
 
     private HttpSession getSession() {
@@ -54,7 +44,7 @@ public class KakaoOAuthService implements OAuthService, KakaoMessageService {
         getSession().setAttribute("access_token", accessToken);
     }
 
-    private String getAccessToken() {
+    public String provideAccessToken() {
         return (String) getSession().getAttribute("access_token");
     }
 
@@ -104,7 +94,7 @@ public class KakaoOAuthService implements OAuthService, KakaoMessageService {
             KakaoUserProfile response = restClient.get()
                 .uri(url)
                 .headers(headers -> {
-                    headers.setBearerAuth(getAccessToken());
+                    headers.setBearerAuth(provideAccessToken());
                     headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
                 })
                 .retrieve()
@@ -117,43 +107,4 @@ public class KakaoOAuthService implements OAuthService, KakaoMessageService {
             throw new RuntimeException("사용자 정보 요청 실패: " + e.getMessage(), e);
         }
     }
-
-    @Override
-    public String createTextMessage(String userMessage) {
-        String message =
-            (userMessage != null && !userMessage.isBlank()) ? userMessage : "기본 메시지입니다.";
-
-        return """
-                {
-                    "object_type": "text",
-                    "text": "%s",
-                    "link": {
-                        "web_url": "http://localhost:8080",
-                        "mobile_web_url": "http://localhost:8080"
-                    }
-                }
-            """.formatted(message);
-    }
-
-    @Override
-    public void sendTextMessage(String templateJson) {
-        var url = kapiHost + "/v2/api/talk/memo/default/send";
-
-        MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-        body.add("template_object", templateJson);
-
-        try {
-            restClient.post()
-                .uri(url)
-                .headers(headers -> headers.setBearerAuth(getAccessToken()))
-                .contentType(MediaType.APPLICATION_FORM_URLENCODED)
-                .body(body)
-                .retrieve()
-                .body(Void.class);
-
-        } catch (RestClientException e) {
-            throw new RuntimeException("카카오 메시지 전송 중 오류 발생");
-        }
-    }
-
 }
